Refactor #1 move everything that has to do with building the validator out to builder/
